Update requirements.txt
Here is the requirements.txt file. This file contains the steps to create a requirements.txt file that lists all the Python packages that are required to run the YouTube to Spotify playlist automation program, you will need to install the required packages and then run the pip freeze command. I also listed the packages that are suppose to be in the requirements.txt file.